pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "polymarket-trades"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "psycopg2-binary>=2.9",
]

[project.scripts]
polymarket-trades = "polymarket_trades.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
